Disable snapshot save button when save isn't possible
