+++ frontend/packages/kubevirt-plugin/integration-tests/deploy-kubevirt-gating.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+export ON_CI="ON_CI"
+
 KUBEVIRT_VERSION=$(curl -s https://github.com/kubevirt/kubevirt/releases/latest | grep -o "v[0-9]\.[0-9]*\.[0-9]*")
 
 VIRTCTL_DOWNLOAD_URL="https://github.com/kubevirt/kubevirt/releases/download/${KUBEVIRT_VERSION}/virtctl-${KUBEVIRT_VERSION}"
