+++ frontend/packages/kubevirt-plugin/locales/en/kubevirt-plugin.json
@@ -546,19 +546,22 @@
   "Tolerations has been updated outside this flow.": "Tolerations has been updated outside this flow.",
   "Saving these changes will override any Tolerations previously saved.": "Saving these changes will override any Tolerations previously saved.",
   "Reload Tolerations": "Reload Tolerations",
+  "No disks found to include in the snapshot": "No disks found to include in the snapshot",
+  "Only disks with a snapshot-supported storage class defined are included in snapshots. No such disks were found.": "Only disks with a snapshot-supported storage class defined are included in snapshots. No such disks were found.",
+  "To take a snapshot you can either edit an existing disk to add a snapshot-supported storage class or add a new disk with a compatible storage class defined. For further details, please contact your cluster admin.": "To take a snapshot you can either edit an existing disk to add a snapshot-supported storage class or add a new disk with a compatible storage class defined. For further details, please contact your cluster admin.",
+  "Learn more about snapshots": "Learn more about snapshots",
   "Take Snapshot": "Take Snapshot",
-  "Snapshot only includes disks backed by a snapshot supported storage class": "Snapshot only includes disks backed by a snapshot supported storage class",
+  "Snapshot only includes disks backed by a snapshot-supported storage class": "Snapshot only includes disks backed by a snapshot-supported storage class",
   "Snapshot Name": "Snapshot Name",
-  "Disks included in this snapshot ({{count}})": "Disks included in this snapshot ({{count}})",
-  "Disks included in this snapshot ({{count}})_plural": "Disks included in this snapshot ({{count}})s",
+  "unsupported approve checkbox": "unsupported approve checkbox",
+  "I am aware of this warning and wish to proceed": "I am aware of this warning and wish to proceed",
+  "Disks included in this snapshot ({{count}})": "Disk included in this snapshot ({{count}})",
+  "Disks included in this snapshot ({{count}})_plural": "Disks included in this snapshot ({{count}})",
   "The following disk will not be included in the snapshot": "The following disk will not be included in the snapshot",
   "The following disk will not be included in the snapshot_plural": "The following disk will not be included in the snapshots",
   "Edit the disk or contact your cluster admin for further details.": "Edit the disk or contact your cluster admin for further details.",
-  "Edit the disk or contact your cluster admin for further details._plural": "Edit the disk or contact your cluster admin for further details.s",
-  "Learn more about snapshots": "Learn more about snapshots",
+  "Edit the disk or contact your cluster admin for further details._plural": "Edit the disks or contact your cluster admin for further details.",
   "The VM {{vmName}} is still running. It will be powered off.": "The VM {{vmName}} is still running. It will be powered off.",
-  "unsupported approve checkbox": "unsupported approve checkbox",
-  "I am aware of this warning and wish to proceed": "I am aware of this warning and wish to proceed",
   "Restore Snapshot": "Restore Snapshot",
   "Are you sure you want to restore <1>{snapshotName}</1> snapshot?": "Are you sure you want to restore <1>{snapshotName}</1> snapshot?",
   "Restore": "Restore",
