+++ frontend/packages/kubevirt-plugin/src/components/modals/snapshot-modal/snapshots-modal.scss
